requirements.txt
multion
python-dotenv
mem0ai
akkio 
pandas 
plotly
mailchimp3
